requirements.txt
APScheduler==3.6.3
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
pycodestyle==2.6.0
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.5
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
